Update setup script

+++ setup.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 
+rm -rf .skel
 git clone https://github.com/tiberiuichim/fullstack-skeleton .skel
 
 if [ -d frontend ]; then
@@ -7,7 +8,13 @@ if [ -d frontend ]; then
   rm -rf .skel/_frontend
 fi;
 
-cp -r .skel/* .
-mv env.example .env
-mv README.md.new README.md
+cp -rna .skel/* .
+
+if [ ! -f .env ]; then
+  mv .skel/tpl/env .env
+fi
+
+if [ ! -f README.md ]; then
+  mv README.md.new README.md
+fi
 rm -rf .skel
